review changes

+++ exchanges/bybit/bybit_ws_futures.go
@@ -314,7 +314,7 @@ func (by *Bybit) wsFuturesHandleData(respRaw []byte) error {
 					AssetType:    a,
 					Side:         oSide,
 					Price:        response.TradeData[i].Price,
-					Amount:       float64(response.TradeData[i].Size),
+					Amount:       response.TradeData[i].Size,
 					Timestamp:    response.TradeData[i].Time,
 				})
 			}
